[KAN-32] Register API (#13)
This api allow users to register with their names, emails and passwords.
Since this is the first version, no additional verification is applied.
It simply checks basic email requirements, and the email domain. That's
it.

The API is located at `/api/register/` and is only valid for `POST`
requests. It accepts a post form with three fields: `name`, `email` and
`password`.

If the register succeed, it returns HTTP 201, else HTTP 400 or 500,
depending on the exception.