Stripe checkout demo work (#16)
This change contains an API prototype that completes the checkout flow with users credit card information, enabling the users to pay money for the stuff they bought on this website. The related keys is configured in a secret file so devs can change it accordingly to suit their need, whether for testing or for real production.

Related APIs are under controller/checkout.py, and in addition to the APIs, an secret provider is implemented to serve as a middle layer between business logic and the underlying key storage logic.