Frontend+backend deployments (#22)
This change enable the FE+BE deployments on Both AWS and Azure with
updated config files and startup files. The necessary settings have
already been updated on AWS.

The frontend code is taken into this repo as a submodule. After
completing a feature the dev need to update the refs of the submodule so
the backend can correctly consume the changes. This submodule is
required unless the frontend set an dedicated deployment pipeline and a
mechanism to control feature synchronism with the backend code, or merge
two repo into one mono-repo.

Also tweaked gitignore files so people on Mac won't commit `.DS_Store`
files into the repo.